Search: News Result
Plus generic Result code reorganization
