+++ ReactNative/js/screens/SearchResults/components/results/GenericResult/snippets/SnippetList.jsx
@@ -11,8 +11,8 @@ import React from 'react';
 import { View, Text, TouchableWithoutFeedback, StyleSheet } from 'react-native';
 import { getMessage } from 'browser-core-user-agent-ios/build/modules/core/i18n';
 
-import { withTheme } from '../../../contexts/theme';
-import NativeDrawable from '../../../components/NativeDrawable';
+import { withTheme } from '../../../../../../contexts/theme';
+import NativeDrawable from '../../../../../../components/NativeDrawable';
 
 const getStyles = theme =>
   StyleSheet.create({
