+++ package.json
@@ -19,6 +19,7 @@
   "dependencies": {
     "@babel/core": "^7.7.2",
     "@cliqz/component-ui-logo": "^0.4.4",
+    "@cliqz/component-ui-snippet-news": "^0.4.0",
     "@cliqz/component-ui-snippet-weather": "^0.4.5",
     "@cliqz/indexeddbshim": "^4.1.2",
     "@react-native-community/netinfo": "^4.6.1",
